server/requirements.txt
websockets>=12.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.53.0
python-dotenv>=1.0.0
google-adk>=0.1.6
google-cloud-discoveryengine>=0.12.0
google-cloud-storage>=2.10.0